+++ webfrontend/htmlauth/system/tools/linfo/src/Linfo/Common.php
@@ -25,18 +25,20 @@
 
 namespace Linfo;
 
+use Linfo\Parsers\IO;
+
 class Common
 {
     protected static $settings = [],
         $lang = [];
 
-    // Used for unit tests
-    public static $path_prefix = false;
+    public static $io;
 
-    public static function config(Linfo $linfo)
+    public static function config(Linfo $linfo, IO $io)
     {
         self::$settings = $linfo->getSettings();
         self::$lang = $linfo->getLang();
+        self::$io = $io;
     }
 
     public static function unconfig()
@@ -76,11 +78,11 @@ class Common
     // Get a file who's contents should just be an int. Returns zero on failure.
     public static function getIntFromFile($file)
     {
-        return self::getContents($file, 0);
+        return (int) self::getContents($file, 0);
     }
 
     // Convert bytes to stuff like KB MB GB TB etc
-    public static function byteConvert($size, $precision = 2)
+    public static function byteConvert($size, $precision = 2, $notation = null, $bits = false)
     {
 
         // Sanity check
@@ -89,12 +91,20 @@ class Common
         }
 
         // Get the notation
-        $notation = self::$settings['byte_notation'] == 1000 ? 1000 : 1024;
+        if ($notation === null) {
+            $notation = self::$settings['byte_notation'] == 1000 ? 1000 : 1024;
+        }
 
         // Fixes large disk size overflow issue
         // Found at http://www.php.net/manual/en/function.disk-free-space.php#81207
-        $types = array('B', 'KB', 'MB', 'GB', 'TB');
-        $types_i = array('B', 'KiB', 'MiB', 'GiB', 'TiB');
+        if ($bits) {
+            $types = array('b', 'Kb', 'Mb', 'Gb', 'Tb');
+            $types_i = array('b', 'Kib', 'Mib', 'Gib', 'Tib');
+        } else {
+            $types = array('B', 'KB', 'MB', 'GB', 'TB');
+            $types_i = array('B', 'KiB', 'MiB', 'GiB', 'TiB');
+        }
+
         for ($i = 0; $size >= $notation && $i < (count($types) - 1); $size /= $notation, $i++);
 
         return(round($size, $precision).' '.($notation == 1000 ? $types[$i] : $types_i[$i]));
@@ -135,7 +145,7 @@ class Common
         }
 
         if ($seconds > 0) {
-            $return[] = $seconds.(date('m/d') == '06/03' ? ' sex' : ' '.self::$lang['seconds']);
+            $return[] = $seconds.' '.self::$lang['seconds'];
         }
 
         return implode(', ', $return);
@@ -144,23 +154,13 @@ class Common
     // Get a file's contents, or default to second param
     public static function getContents($file, $default = '')
     {
-        if (is_string(self::$path_prefix)) {
-            $file = self::$path_prefix.$file;
-        }
-
-        return !is_file($file) || !is_readable($file) || !($contents = @file_get_contents($file)) ? $default : trim($contents);
+        return self::$io->getContents($file, $default);
     }
 
     // Like above, but in lines instead of a big string
     public static function getLines($file)
     {
-        return !is_file($file) || !is_readable($file) || !($lines = @file($file, FILE_SKIP_EMPTY_LINES)) ? [] : $lines;
-    }
-
-    // Make a string safe for being in an xml tag name
-    public static function xmlStringSanitize($string)
-    {
-        return strtolower(preg_replace('/([^a-zA-Z]+)/', '_', $string));
+        return self::$io->getLines($file);
     }
 
     // Get a variable from a file. Include it in this function to avoid
@@ -193,5 +193,26 @@ class Common
         }
 
         return count(array_intersect($needles, $haystack)) > 0;
+    }
+
+    // Easily determine if a string contains any of the snippets
+    public static function anyInString($string, $snippets)
+    {
+	foreach ($snippets as $snippet) {
+	    if (strpos($string, $snippet) !== false) {
+                 return true;
+	    }
+	}
+	return false;
+    }
+
+    // Given a string that should contain a float or a decimal, get back either
+    public static function strToInt($str)
+    {
+        if (!is_string($str))
+            return $str;
+        if (strpos($str, '.') !== false)
+            return (float) $str;
+        return (int) $str;
     }
 }