+++ webfrontend/htmlauth/system/tools/linfo/src/Linfo/Linfo.php
@@ -26,6 +26,7 @@
 namespace Linfo;
 
 use Linfo\Parsers\CallExt;
+use Linfo\Parsers\FileIO;
 use Linfo\Exceptions\FatalException;
 use Linfo\Meta\Errors;
 use ReflectionClass;
@@ -63,7 +64,7 @@ class Linfo
         $this->linfo_localdir = dirname(dirname(__DIR__)).'/';
 
         // Get our version from git setattribs
-        $scm = '2019-10-09 09:10:15 -0700';
+        $scm = '2025-11-10 20:37:52 -0800';
         list($this->version) = strpos($scm, '$') !== false ? array('git') : explode(' ', $scm);
 
         // Run through dependencies / sanity checking
@@ -80,8 +81,11 @@ class Linfo
         $this->loadSettings($settings);
         $this->loadLanguage();
 
+        // We'll be hitting local files on disk in their usual paths
+        $io = new FileIO(false);
+
         // Some classes need our vars; config them
-        Common::config($this);
+        Common::config($this, $io);
         CallExt::config($this);
 
         // Determine OS
@@ -108,185 +112,181 @@ class Linfo
     // Load everything, while obeying permissions...
     public function scan()
     {
-        $reflector = new ReflectionClass($this->parser);
-
         // Prime parser. Do things not appropriate to do in constructor. Most OS classes
         // don't have this.
-        if ($reflector->hasMethod('init') && ($method = $reflector->getMethod('init'))) {
-            $method->invoke($this->parser);
-        }
+        $this->parser->init();
 
         // Array fields, tied to method names and default values...
-        $fields = array(
-            'OS' => array(
+        $fields = [
+            'OS' => [
                 'show' => !empty($this->settings['show']['os']),
                 'default' => '',
                 'method' => 'getOS',
-            ),
+            ],
 
-            'Kernel' => array(
+            'Kernel' => [
                 'show' => !empty($this->settings['show']['kernel']),
                 'default' => '',
                 'method' => 'getKernel',
-            ),
+            ],
 
-            'AccessedIP' => array(
+            'AccessedIP' => [
                 'show' => !isset($this->settings['show']['ip']) || !empty($this->settings['show']['ip']),
                 'default' => '',
                 'method' => 'getAccessedIP',
-            ),
+            ],
 
-            'Distro' => array(
+            'Distro' => [
                 'show' => !empty($this->settings['show']['distro']),
                 'default' => '',
                 'method' => 'getDistro',
-            ),
+            ],
 
-            'RAM' => array(
+            'RAM' => [
                 'show' => !empty($this->settings['show']['ram']),
                 'default' => [],
                 'method' => 'getRam',
-            ),
+            ],
 
-            'HD' => array(
+            'HD' => [
                 'show' => !empty($this->settings['show']['hd']),
                 'default' => [],
                 'method' => 'getHD',
-            ),
+            ],
 
-            'Mounts' => array(
+            'Mounts' => [
                 'show' => !empty($this->settings['show']['mounts']),
                 'default' => [],
                 'method' => 'getMounts',
-            ),
+            ],
 
-            'Load' => array(
+            'Load' => [
                 'show' => !empty($this->settings['show']['load']),
                 'default' => [],
                 'method' => 'getLoad',
-            ),
+            ],
 
-            'HostName' => array(
+            'HostName' => [
                 'show' => !empty($this->settings['show']['hostname']),
                 'default' => '',
                 'method' => 'getHostName',
-            ),
+            ],
 
-            'UpTime' => array(
+            'UpTime' => [
                 'show' => !empty($this->settings['show']['uptime']),
                 'default' => [],
                 'method' => 'getUpTime',
-            ),
+            ],
 
-            'CPU' => array(
+            'CPU' => [
                 'show' => !empty($this->settings['show']['cpu']),
                 'default' => [],
                 'method' => 'getCPU',
-            ),
+            ],
 
-            'Model' => array(
+            'Model' => [
                 'show' => !empty($this->settings['show']['model']),
                 'default' => [],
                 'method' => 'getModel',
-            ),
+            ],
 
-            'CPUArchitecture' => array(
+            'CPUArchitecture' => [
                 'show' => !empty($this->settings['show']['cpu']),
                 'default' => '',
                 'method' => 'getCPUArchitecture',
-            ),
+            ],
 
-            'Network Devices' => array(
+            'Network Devices' => [
                 'show' => !empty($this->settings['show']['network']),
                 'default' => [],
                 'method' => 'getNet',
-            ),
+            ],
 
-            'Devices' => array(
+            'Devices' => [
                 'show' => !empty($this->settings['show']['devices']),
                 'default' => [],
                 'method' => 'getDevs',
-            ),
+            ],
 
-            'Temps' => array(
+            'Temps' => [
                 'show' => !empty($this->settings['show']['temps']),
                 'default' => [],
                 'method' => 'getTemps',
-            ),
+            ],
 
-            'Battery' => array(
+            'Battery' => [
                 'show' => !empty($this->settings['show']['battery']),
                 'default' => [],
                 'method' => 'getBattery',
-            ),
+            ],
 
-            'Raid' => array(
+            'Raid' => [
                 'show' => !empty($this->settings['show']['raid']),
                 'default' => [],
                 'method' => 'getRAID',
-            ),
+            ],
 
-            'Wifi' => array(
+            'Wifi' => [
                 'show' => !empty($this->settings['show']['wifi']),
                 'default' => [],
                 'method' => 'getWifi',
-            ),
+            ],
 
-            'SoundCards' => array(
+            'SoundCards' => [
                 'show' => !empty($this->settings['show']['sound']),
                 'default' => [],
                 'method' => 'getSoundCards',
-            ),
+            ],
 
-            'processStats' => array(
+            'processStats' => [
                 'show' => !empty($this->settings['show']['process_stats']),
                 'default' => [],
                 'method' => 'getProcessStats',
-            ),
+            ],
 
-            'services' => array(
+            'services' => [
                 'show' => !empty($this->settings['show']['services']),
                 'default' => [],
                 'method' => 'getServices',
-            ),
+            ],
 
-            'numLoggedIn' => array(
+            'numLoggedIn' => [
                 'show' => !empty($this->settings['show']['numLoggedIn']),
                 'default' => false,
                 'method' => 'getnumLoggedIn',
-            ),
+            ],
 
-            'virtualization' => array(
+            'virtualization' => [
                 'show' => !empty($this->settings['show']['virtualization']),
                 'default' => [],
                 'method' => 'getVirtualization',
-            ),
+            ],
 
-            'cpuUsage' => array(
+            'cpuUsage' => [
                 'show' => !empty($this->settings['cpu_usage']),
                 'default' => false,
                 'method' => 'getCPUUsage',
-            ),
+            ],
 
-            'phpVersion' => array(
+            'phpVersion' => [
                 'show' => !empty($this->settings['show']['phpversion']),
                 'default' => false,
                 'method' => 'getPhpVersion',
-            ),
+            ],
 
-            'webService' => array(
+            'webService' => [
                 'show' => !empty($this->settings['show']['webservice']),
                 'default' => false,
                 'method' => 'getWebService',
-            ),
+            ],
 
             // Extra info such as which fields to not show
-            'contains' => array(
+            'contains' => [
                 'show' => true,
                 'default' => [],
                 'method' => 'getContains',
-            ),
-        );
+            ],
+        ];
 
         foreach ($fields as $key => $data) {
             if (!$data['show']) {
@@ -294,10 +294,11 @@ class Linfo
                 continue;
             }
 
-            try {
-                $method = $reflector->getMethod($data['method']);
-                $this->info[$key] = $method->invoke($this->parser);
-            } catch (ReflectionException $e) {
+            if (method_exists($this->parser, $data['method'])) {
+                $this->info[$key] = call_user_func_array(
+                    [$this->parser, $data['method']],
+                    isset($data['args']) ? $data['args'] : []);
+            } else {
                 $this->info[$key] = $data['default'];
             }
         }
@@ -354,6 +355,12 @@ class Linfo
         if (!is_file($this->linfo_localdir.'src/Linfo/Lang/'.$settings['language'].'.php')) {
             $settings['language'] = 'en';
         }
+
+        // Make sure we have show, as lots of stuff assumes it exists
+        $settings['show'] = isset($settings['show']) ? $settings['show'] : [];
+
+        // And timer, as this is referenced in most OS methods
+        $settings['timer'] = isset($settings['timer']) ? $settings['timer'] : false;
 
         $this->settings = $settings;
     }