+++ webfrontend/htmlauth/system/tools/linfo/src/Linfo/OS/Linux.php
@@ -334,10 +334,10 @@ class Linux extends Unixcommon
         }
 
         // Seconds
-        list($seconds) = explode(' ', $contents, 1);
+        list($seconds,) = explode(' ', $contents);
 
         // Get it textual, as in days/minutes/hours/etc
-        $uptime = Common::secondsConvert(ceil($seconds));
+        $uptime = Common::secondsConvert(ceil((float)$seconds));
 
         // Now find out when the system was booted
         $contents = Common::getContents('/proc/stat', false);
@@ -377,7 +377,7 @@ class Linux extends Unixcommon
         // Get partitions
         $partitions = [];
         $partitions_contents = Common::getContents('/proc/partitions');
-        if (@preg_match_all('/(\d+)\s+([a-z]{3})(\d+)$/m', $partitions_contents, $partitions_match, PREG_SET_ORDER) > 0) {
+        if (@preg_match_all('/(\d+)\s+([a-z]{3}|nvme\d+n\d+|[a-z]+\d+)(p?\d+)$/m', $partitions_contents, $partitions_match, PREG_SET_ORDER) > 0) {
             // Go through each match
             foreach ($partitions_match as $partition) {
                 $partitions[$partition[2]][] = array(
@@ -391,7 +391,7 @@ class Linux extends Unixcommon
         $drives = [];
 
         // Get actual drives
-        foreach ((array) @glob('/sys/block/*/device/model', GLOB_NOSORT) as $path) {
+        foreach ((array) @glob('/sys/block/*/device/uevent', GLOB_NOSORT) as $path) {
 
             // Parts of the path
             $parts = explode('/', $path);
@@ -406,9 +406,19 @@ class Linux extends Unixcommon
                 list(, $reads, $writes) = $statMatches;
             }
 
+            $type = '';
+
+            if (Common::getContents(dirname(dirname($path)).'/queue/rotational') == 0) {
+                if (Common::getContents(dirname($path).'/type') == 'SD') {
+                    $type = ' (SD)';
+                } else {
+                    $type = ' (SSD)';
+                }
+            }
+
             // Append this drive on
             $drives[] = array(
-                'name' => Common::getContents($path, 'Unknown').(Common::getContents(dirname(dirname($path)).'/queue/rotational') == 0 ? ' (SSD)' : ''),
+                'name' => Common::getContents(dirname($path).'/model', 'Unknown').$type,
                 'vendor' => Common::getContents(dirname($path).'/vendor', 'Unknown'),
                 'device' => '/dev/'.$parts[3],
                 'reads' => $reads,
@@ -523,15 +533,23 @@ class Linux extends Unixcommon
             // Wacky location
             foreach ((array) @glob('/sys/class/hwmon/hwmon*/{,device/}*_input', GLOB_NOSORT | GLOB_BRACE) as $path) {
                 $initpath = rtrim($path, 'input');
-                $value = Common::getContents($path);
+                $value = Common::getIntFromFile($path);
                 $base = basename($path);
                 $labelpath = $initpath.'label';
+                $modelpath = dirname($path).'/device/model';
                 $showemptyfans = isset($this->settings['temps_show0rpmfans']) ? $this->settings['temps_show0rpmfans'] : false;
-                $drivername = @basename(@readlink(dirname($path).'/driver')) ?: false;
+                $drivername = basename(@readlink(dirname($path).'/driver')) ?: false;
 
                 // Temperatures
                 if (is_file($labelpath) && strpos($base, 'temp') === 0) {
                     $label = Common::getContents($labelpath);
+                    $value /= $value > 10000 ? 1000 : 1;
+                    $unit = 'C'; // I don't think this is ever going to be in F
+                }
+
+                // Devices (such as hard drives)
+                else if (is_file($modelpath) && strpos($base, 'temp') === 0) {
+                    $label = Common::getContents($modelpath);
                     $value /= $value > 10000 ? 1000 : 1;
                     $unit = 'C'; // I don't think this is ever going to be in F
                 }
@@ -648,11 +666,13 @@ class Linux extends Unixcommon
         // Can't?
         if ($contents == false) {
             Errors::add('Linfo Core', '/proc/mounts does not exist');
+            return [];
         }
 
         // Parse
         if (@preg_match_all('/^(\S+) (\S+) (\S+) (.+) \d \d$/m', $contents, $match, PREG_SET_ORDER) === false) {
             Errors::add('Linfo Core', 'Error parsing /proc/mounts');
+            return [];
         }
 
         // Return these
@@ -679,9 +699,13 @@ class Linux extends Unixcommon
             $mount[2] = stripcslashes($mount[2]);
 
             // Get these
-            $size = @disk_total_space($mount[2]);
-            $free = @disk_free_space($mount[2]);
-            $used = $size != false && $free != false ? $size - $free : false;
+            if (is_readable($mount[2])) {
+                $size = disk_total_space($mount[2]);
+                $free = disk_free_space($mount[2]);
+                $used = $size != false && $free != false ? $size - $free : false;
+            }else{
+                $size = $free = $used = false;
+            }
 
             // If it's a symlink, find out where it really goes.
             // (using realpath instead of readlink because the former gives absolute paths)
@@ -692,7 +716,7 @@ class Linux extends Unixcommon
             }
 
             // Optionally get mount options
-            if ($this->settings['show']['mounts_options'] && !in_array($mount[3], (array) $this->settings['hide']['fs_mount_options'])) {
+            if (isset($this->settings['show']['mounts_options']) && $this->settings['show']['mounts_options'] && !in_array($mount[3], (array) $this->settings['hide']['fs_mount_options'])) {
                 $mount_options = explode(',', $mount[4]);
             } else {
                 $mount_options = [];
@@ -746,8 +770,7 @@ class Linux extends Unixcommon
         $usb_ids || Errors::add('Linux Device Finder', 'Cannot find usb.ids; ensure usbutils is installed.');
 
         // Class that does it
-        $hw = new Hwpci($usb_ids, $pci_ids);
-        $hw->work('linux');
+        $hw = new Hwpci($usb_ids, $pci_ids, 'linux', true);
 
         return $hw->result();
     }
@@ -768,6 +791,10 @@ class Linux extends Unixcommon
         // Store it here
         $raidinfo = [];
 
+        if (!isset($this->settings['raid'])) {
+          return $raidinfo;
+        }
+
         // mdadm?
         if (array_key_exists('mdadm', (array) $this->settings['raid']) && !empty($this->settings['raid']['mdadm'])) {
 
@@ -780,7 +807,7 @@ class Linux extends Unixcommon
             }
 
             // Parse
-            @preg_match_all('/(\S+)\s*:\s*(\w+)\s*raid(\d+)\s*([\w+\[\d+\] (\(\w\))?]+)\n\s+(\d+) blocks[^[]+\[(\d\/\d)\] \[([U\_]+)\]/mi', (string) $mdadm_contents, $match, PREG_SET_ORDER);
+            @preg_match_all('/(?P<device>\S+)\s*:\s*(?P<state>\w+)(?P<blurb>\s+\([^)]+\))?\s*raid(?P<level>\d+)\s*(?P<drives>[\w+\[\d+\] (\(\w\))?]+)\n\s+(?P<blocks>\d+) blocks[^[]+\[(?P<counts>\d\/\d)\] \[(?P<chart>[U\_]+)\]/mi', (string) $mdadm_contents, $match, PREG_SET_ORDER);
 
             // Store them here
             $mdadm_arrays = [];
@@ -792,14 +819,14 @@ class Linux extends Unixcommon
                 $drives = [];
 
                 // Parse drives
-                foreach (explode(' ', $array[4]) as $drive) {
+                foreach (explode(' ', $array['drives']) as $drive) {
 
                     // Parse?
-                    if (preg_match('/([\w\d]+)\[\d+\](\(\w\))?/', $drive, $match_drive) == 1) {
+                    if (preg_match('/(?P<device>[\w\d]+)\[\d+\](?P<state>\(\w\))?/', $drive, $match_drive) == 1) {
 
                         // Determine a status other than normal, like if it failed or is a spare
-                        if (array_key_exists(2, $match_drive)) {
-                            switch ($match_drive[2]) {
+                        if (isset($match_drive['state'])) {
+                            switch ($match_drive['state']) {
                                 case '(S)':
                                     $drive_state = 'spare';
                                 break;
@@ -821,21 +848,27 @@ class Linux extends Unixcommon
 
                         // Append this drive to the temp drives array
                         $drives[] = array(
-                            'drive' => '/dev/'.$match_drive[1],
+                            'drive' => '/dev/'.$match_drive['device'],
                             'state' => $drive_state,
                         );
                     }
                 }
 
+                $state = $array['state'];
+
+                if(isset($array['blurb'])) {
+                  $state .= $array['blurb'];
+                }
+
                 // Add record of this array to arrays list
                 $mdadm_arrays[] = array(
-                    'device' => '/dev/'.$array[1],
-                    'status' => $array[2],
-                    'level' => $array[3],
+                    'device' => '/dev/'.$array['device'],
+                    'status' => $state,
+                    'level' => $array['level'],
                     'drives' => $drives,
-                    'size' => Common::byteConvert($array[5] * 1024),
-                    'count' => $array[6],
-                    'chart' => $array[7],
+                    'size' => Common::byteConvert($array['blocks'] * 1024),
+                    'count' => $array['counts'],
+                    'chart' => $array['chart'],
                 );
             }
 
@@ -943,12 +976,32 @@ class Linux extends Unixcommon
                 $type_contents = strtoupper(Common::getContents($path.'/device/modalias'));
                 list($type_match) = explode(':', $type_contents, 2);
 
-                if (in_array($type_match, array('PCI', 'USB'))) {
+		if(is_readable($path.'/uevent')){
+                    $uevent_contents = @parse_ini_file($path.'/uevent');
+                } else{
+                    $uevent_contents = false;
+                }
+
+                if(is_readable($path.'/device/uevent')){
+                    $device_uevent_contents = @parse_ini_file($path.'/device/uevent');
+                } else{
+                    $device_uevent_contents = false;
+                }
+
+                if ($uevent_contents != false && isset($uevent_contents['DEVTYPE'])) {
+                  $type = ucfirst($uevent_contents['DEVTYPE']);
+                    if (in_array($type_match, array('PCI', 'USB'))){
+                        $type .= ' ('.$type_match.')';
+                    }
+                    if ($device_uevent_contents != false && isset($device_uevent_contents['DRIVER'])) {
+                        $type .= ' ('.$device_uevent_contents['DRIVER'].')';
+                    }
+                } elseif (in_array($type_match, array('PCI', 'USB'))) {
                     $type = 'Ethernet ('.$type_match.')';
 
                     // Driver maybe?
-                    if (($uevent_contents = @parse_ini_file($path.'/device/uevent')) && isset($uevent_contents['DRIVER'])) {
-                        $type .= ' ('.$uevent_contents['DRIVER'].')';
+                    if ($device_uevent_contents != false && isset($device_uevent_contents['DRIVER'])) {
+                        $type .= ' ('.$device_uevent_contents['DRIVER'].')';
                     }
                 } elseif ($type_match == 'VIRTIO') {
                     $type = 'VirtIO';
@@ -985,7 +1038,7 @@ class Linux extends Unixcommon
                 // These were determined above
                 'state' => $state,
                 'type' => $type ?: 'N/A',
-                'port_speed' => $speed > 0 ? $speed : false,
+                'port_speed' => $speed > 0 ? $speed * 1000 * 1000 : false,
             );
         }
 
@@ -1188,7 +1241,9 @@ class Linux extends Unixcommon
             $status_contents = Common::getContents($process);
 
             // Try getting state
-            @preg_match('/^State:\s+(\w)/m', $status_contents, $state_match);
+	    if (!@preg_match('/^State:\s+(\w)/m', $status_contents, $state_match)) {
+                continue;
+            }
 
             // Well? Determine state
             switch ($state_match[1]) {
@@ -1208,7 +1263,9 @@ class Linux extends Unixcommon
             }
 
             // Try getting number of threads
-            @preg_match('/^Threads:\s+(\d+)/m', $status_contents, $threads_match);
+            if (!@preg_match('/^Threads:\s+(\d+)/m', $status_contents, $threads_match)){
+                continue;
+            }
 
             // Well?
             if ($threads_match) {
@@ -1403,6 +1460,17 @@ class Linux extends Unixcommon
         // - And even also supports empty files, and just uses said file to identify the distro and ignore version
 
         $contents_distros = array(
+            // Various redhat flavors/derivs
+            array(
+                'file' => '/etc/fedora-release',
+                'regex' => '/^Fedora(?: Core)? release (?P<version>\d+) \((?P<codename>[^)]+)\)$/',
+                'distro' => 'Fedora',
+            ),
+            array(
+                'file' => '/etc/oracle-release',
+                'regex' => '/^Oracle Linux Server release (?P<version>[\d\.]+)/',
+                'distro' => 'Oracle',
+            ),
             array(
                 'file' => '/etc/redhat-release',
                 'regex' => '/^CentOS.+release (?P<version>[\d\.]+) \((?P<codename>[^)]+)\)$/i',
@@ -1413,6 +1481,8 @@ class Linux extends Unixcommon
                 'regex' => '/^Red Hat.+release (?P<version>\S+) \((?P<codename>[^)]+)\)$/i',
                 'distro' => 'RedHat',
             ),
+
+            // Should catch most distros (Ubuntu/etc)
             array(
                 'file' => '/etc/lsb-release',
                 'closure' => function ($ini) {
@@ -1437,11 +1507,7 @@ class Linux extends Unixcommon
                     ) : false;
                  },
             ),
-            array(
-                'file' => '/etc/fedora-release',
-                'regex' => '/^Fedora(?: Core)? release (?P<version>\d+) \((?P<codename>[^)]+)\)$/',
-                'distro' => 'Fedora',
-            ),
+
             array(
                 'file' => '/etc/gentoo-release',
                 'regex' => '/(?P<version>[\d\.]+)$/',
@@ -1482,7 +1548,7 @@ class Linux extends Unixcommon
                     'name' => $distro['distro'],
                     'version' => $info['version'].(isset($info['codename']) ? ' ('.ucfirst($info['codename']).')' : ''),
                 );
-            } elseif (isset($distro['distro'])) {
+            } elseif (isset($distro['distro']) && !isset($distro['regex'])) {
                 return array(
                     'name' => $distro['distro'],
                     'version' => $contents,
@@ -1539,7 +1605,7 @@ class Linux extends Unixcommon
         foreach ($procs as $proc) {
 
             // Does the process match a popular shell, such as bash, csh, etc?
-            if (preg_match('/(?:bash|csh|zsh|ksh)$/', Common::getContents($proc, ''))) {
+            if (Common::anyInString(Common::getContents($proc, ''), ['bash', 'csh', 'zsh', 'ksh'])) {
 
                 // Who owns it, anyway?
                 $owner = fileowner(dirname($proc));
@@ -1584,12 +1650,19 @@ class Linux extends Unixcommon
             return array('type' => 'guest', 'method' => 'OpenVZ');
         }
 
+        $bios_vendor = Common::getContents('/sys/devices/virtual/dmi/id/bios_vendor');
+
         // Veertu guest?
-        if (Common::getContents('/sys/devices/virtual/dmi/id/bios_vendor') == 'Veertu') {
+        if ($bios_vendor == 'Veertu') {
             return array('type' => 'guest', 'method' => 'Veertu');
         }
 
-	// LXC guest?
+        // Parallels guest?
+        if (strpos($bios_vendor, 'Parallels') === 0) {
+            return array('type' => 'guest', 'method' => 'Parallels');
+        }
+
+        // LXC guest?
         if (strpos(Common::getContents('/proc/mounts'), 'lxcfs /proc/') !== false) {
             return array('type' => 'guest', 'method' => 'LXC');
         }