+++ webfrontend/htmlauth/system/tools/linfo/src/Linfo/Output/Html.php
@@ -198,14 +198,14 @@ class Html implements Output
 	<link rel="stylesheet" type="text/css" href="./layout/mobile.css" media="screen and (max-width: 640px)">
 </head>
 <body id="info">
-<!--<div id="header">';
+<div id="header">';
 
   if ($settings['allow_changing_themes']) {
   echo '
   <div id="themeChanger">Theme: <select id="themeChangerSelect">';
 
-    foreach ($allowed_themes as $theme_option)
-      echo '<option'.($theme_option == $chosen_theme ? ' selected' : '').'>'.$theme_option.'</option>';
+    foreach ($allowed_themes as $theme)
+      echo '<option'.($theme == $chosen_theme ? ' selected' : '').' value="'.$theme.'">'.implode(' ', array_map('\ucfirst', explode('_', $theme))).'</option>';
 
   echo '</select>
   </div>
@@ -215,8 +215,8 @@ class Html implements Output
 echo '
 	<h1>'.$info['HostName'].'</h1>
 	<div class="subtitle">'.$lang['header'].'</div>
-</div>-->
-<!--<div class="col2">-->
+</div>
+<div class="col2">
 	<div class="col">
 		<div class="infoTable">
 			<h2>'.$lang['core'].'</h2>
@@ -266,7 +266,7 @@ echo '
     }
 
     // Uptime
-    if (!empty($settings['show']['uptime']) && $info['UpTime']) {
+    if (!empty($settings['show']['uptime']) && $info['UpTime'] && is_array($info['UpTime'])) {
         $core[] = array($lang['uptime'],
             $info['UpTime']['text'].
                 (isset($info['UpTime']['bootedTimestamp']) && $info['UpTime']['bootedTimestamp'] ? '; booted '.date($settings['dates'], $info['UpTime']['bootedTimestamp']) : ''), );
@@ -394,7 +394,7 @@ echo '
 					<td>'.Common::byteConvert($info['RAM']['total']).'</td>
 					<td>'.Common::byteConvert($info['RAM']['total'] - $info['RAM']['free']).'</td>
 					<td>'.Common::byteConvert($info['RAM']['free']).'</td>
-					<td>'.self::generateBarChart(round(($info['RAM']['total'] - $info['RAM']['free']) * 100 / $info['RAM']['total'])).'</td>
+					<td>'.($info['RAM']['total'] > 0 ? self::generateBarChart(round(($info['RAM']['total'] - $info['RAM']['free']) * 100 / $info['RAM']['total'])) : '').'</td>
 				</tr>';
         $have_swap = (isset($info['RAM']['swapFree']) || isset($info['RAM']['swapTotal']));
         if ($have_swap && !empty($info['RAM']['swapTotal'])) {
@@ -472,7 +472,7 @@ echo '
 				<tr>
 					<td>'.$device.'</td>'.($show_type ? '
 					<td>'.$stats['type'].'</td>' : '').($show_speed ? '
-					<td>'.(isset($stats['port_speed']) && $stats['port_speed'] !== false ? $stats['port_speed'].'Mb/s' : '').'</td>' : '').'
+					<td>'.(isset($stats['port_speed']) && $stats['port_speed'] !== false ? Common::byteConvert($stats['port_speed'], 2, 1000, true).'/s' : '').'</td>' : '').'
 					<td>'.Common::byteConvert($stats['sent']['bytes']).'</td>
 					<td>'.Common::byteConvert($stats['recieved']['bytes']).'</td>
 					<td class="net_'.$stats['state'].'">'.ucfirst($stats['state']).'</td>
@@ -598,8 +598,8 @@ echo '
     }
 
         echo '
-	<!--</div>
-	<div class="col">-->';
+	</div>
+	<div class="col">';
 
     // Show hardware?
     if (!empty($settings['show']['devices'])) {
@@ -619,12 +619,14 @@ echo '
 				';
         $num_devs = count($info['Devices']);
         if ($num_devs > 0) {
-            for ($i = 0; $i < $num_devs; ++$i) {
+            foreach($info['Devices'] as $device) {
                 echo '
 				<tr>
-					<td class="center">'.$info['Devices'][$i]['type'].'</td>
-					',$show_vendor ? '<td>'.($info['Devices'][$i]['vendor'] ? $info['Devices'][$i]['vendor'] : 'Unknown').'</td>' : '','
-					<td>'.$info['Devices'][$i]['device'].'</td>
+					<td class="center">'.$device['type'].'</td>
+					',$show_vendor ? '<td>'.($device['vendor'] ? $device['vendor'] : 'Unknown').'</td>' : '','
+					<td>'.$device['device'].(isset($device['count']) && $device['count'] > 1 ? ' <span class="subtitle">(x'.$device['count'].')</span>' : '' ).'
+					'.(isset($device['speed']) ? '<span class="subtitle">('.Common::byteConvert($device['speed'], 2, 1000, true).'/s)</span>' : '').'
+          </td>
 				</tr>';
             }
         } else {
@@ -725,7 +727,7 @@ echo '
 
         echo '
 	</div>
-<!--</div>-->';
+</div>';
 
     // Show file system mounts?
     if (!empty($settings['show']['mounts'])) {